### 2019-01-25 #1202 @whedon
[PRE REVIEW]: surfinpy: A Surface Phase Diagram Generator
**Submitting author:** @symmy596 (<a href="http://orcid.org/0000-0001-6059-497X">Adam Symington</a>)
**Repository:** <a href="https://github.com/symmy596/SurfinPy" target ="_blank">https://github.com/symmy596/SurfinPy</a>
**Version:** 0.8
