### 2019-01-27 #1202 @labarba
👋  @@symmy596 — Doing the pre-checks on your paper, we've found that your references list is missing a bunch of DOIs. Check on the helpful suggestions of our bot ☝️
Can you update your BibTeX entries to include the DOIs?
Also, I request that you don't use double { } around the titles to override the capitalization style of the bibliography. Only use them in acronyms or other individual names that need to be capitalized. Thanks!
