### 2019-01-27 #1202 @labarba
@bjmorgan — Maybe you would like to contribute a review for JOSS on this one?
