### 2019-01-27 #1202 @bjmorgan
@labarba I'd be happy to review, but want to flag a possible [conflict of interest](https://github.com/openjournals/joss/blob/master/COI.md). I'm in the same department as three of the authors on this paper, and have an active collaboration with two of the authors.
