### 2019-01-28 #1202 @mkhorton
I’d be happy to review, I also know @utf and can vouch he’d also be a good
reviewer / appropriate for this.

On Sun, Jan 27, 2019 at 14:07, Benjamin Morgan <notifications@github.com>
wrote:

> @utf <https://github.com/utf> is not on the reviewer list, but has
> published in JOSS and has the right background.
>
