### 2019-01-28 #1210 @whedon
[REVIEW]: surfinpy: A Surface Phase Diagram Generator
**Submitting author:** @symmy596 (<a href="http://orcid.org/0000-0001-6059-497X">Adam Symington</a>)
**Repository:** <a href="https://github.com/symmy596/SurfinPy" target ="_blank">https://github.com/symmy596/SurfinPy</a>
**Version:** v1.0.1
