### 2019-01-28 #1210 @labarba
👋  @symmy596 — Please check your references list, aided by the bot's suggestions in the pre-review issue: https://github.com/openjournals/joss-reviews/issues/1202#issuecomment-457943767
