### 2019-01-28 c13fb23
Updated bib
