+++ paper/paper.bib
@@ -1,50 +1,60 @@
 @article{Canepa2018,
 author = {Canepa, Pieremanuele and Dawson, James A. and Sai Gautam, Gopalakrishnan and Statham, Joel M. and Parker, Stephen C. and Islam, M. Saiful},
-title = {{Particle Morphology and Lithium Segregation to Surfaces of the Li7La3Zr2O12 Solid Electrolyte}},
+title = {Particle Morphology and Lithium Segregation to Surfaces of the Li7La3Zr2O12 Solid Electrolyte},
 journal = {Chemistry of Materials},
 year = {2018},
 volume = {30},
 number = {9},
 pages = {3019-3027},
-        }
+doi = {10.1021/acs.chemmater.8b00649},
+url = {https://pubs.acs.org/doi/10.1021/acs.chemmater.8b00649}
+}
 
 @article{Molinari2012,
 author = {Molinari, Marco and Parker, Stephen C. and Sayle, Dean C. and Islam, M. Saiful},
-title = {{Water Adsorption and Its Effect on the Stability of Low Index Stoichiometric and Reduced Surfaces of Ceria}},
+title = {Water Adsorption and Its Effect on the Stability of Low Index Stoichiometric and Reduced Surfaces of Ceria},
 journal = {The Journal of Physical Chemistry C},
 year = {2012},
 volume = {116},
 number = {12},
 pages = {7073-7082},
-        }
+doi = {10.1021/jp300576b}, 
+url = {https://pubs.acs.org/doi/10.1021/jp300576b}
+}
     
 @article{Tegner2017,
 author = {Tegner, Bengt E. and Molinari, Marco and Kerridge, Andrew and Parker, Stephen C. and Kaltsoyannis, Nikolas},
-title = {{Water Adsorption on AnO2 {111}, {110}, and {100} Surfaces (An = U and Pu): A Density Functional Theory + U Study}},
+title = {Water Adsorption on AnO2 {111}, {110}, and {100} Surfaces (An = U and Pu): A Density Functional Theory + U Study},
 journal = {The Journal of Physical Chemistry C},
 year = {2017},
 volume = {121},
 number = {3},
 pages = {1675-1682},
-        }
+doi = {10.1021/acs.jpcc.6b10986},
+url = {https://pubs.acs.org/doi/10.1021/acs.jpcc.6b10986}
+}
 
 @article{Reuter2003,
 author = {Reuter, Karsten and Scheffler, Matthias},
-title = {{First-Principles Atomistic Thermodynamics for Oxidation Catalysis: Surface Phase Diagrams and Catalytically Interesting Regions}},
+title = {First-Principles Atomistic Thermodynamics for Oxidation Catalysis: Surface Phase Diagrams and Catalytically Interesting Regions},
 journal = {Phys. Rev. Lett.},
 year = {2003},
 volume = {90},
 issue = {4},
 pages = {046103},
-        }
+doi = {10.1103/PhysRevLett.90.046103},
+url = {https://journals.aps.org/prl/abstract/10.1103/PhysRevLett.90.046103}
+}
 
 @article{Marmier2004,
 author = {Marmier, Arnaud and Parker, Stephen C.},
-title = {{Ab initio morphology and surface thermodynamics of $\ensuremath{\alpha}\ensuremath{-}{\mathrm{Al}}_{2}{\mathrm{O}}_{3}$}},
+title = {Ab initio morphology and surface thermodynamics of $\ensuremath{\alpha}\ensuremath{-}{\mathrm{Al}}_{2}{\mathrm{O}}_{3}$},
 journal = {Phys. Rev. B},
 year = {2004},
 volume = {69},
 issue = {11},
[…]
