### 2019-01-28 #1210 @symmy596
Thank you @labarba for the speed that you have gone through this so far. The references should now be updated and the doi's added.
