### 2019-02-08 #1210 @bocklund
@symmy596 looks great! The API and theory is well documented and matplotlib plots look especially nice. I have a few comments:

1. Documentation - Installation instructions: The requirements found in `requirements.txt` should be added to `setup.py`, otherwise users installing from pip will not get all the requirements.
2. Software Paper - References: looks like reference `Reuter2003` did not get correctly inserted into the proof.
3. Functionality - Functionality: Please run the example notebooks and put them on GitHub so the outputs can be rendered. This lets users know that they are getting the right answers when they run the code. The tutorials seem to run fine for me, just not sure if I'm seeing the expected output.
4. HTML for logo is broken for me on [Readthedocs](https://surfinpy.readthedocs.io/en/latest/)
5. Documentation - Community guidelines: Unclear how users should seek support.

Some suggestions (not blocking acceptance, in my opinion)

1. It looks like your tags on GitHub are not updated to the latest release version, but you have used the tags in the past. Something that is nice for contributors and future code-owners is to document the release process steps. Here's an [example of one for ESPEI](https://github.com/PhasesResearchLab/ESPEI/blob/master/RELEASING.rst), which I use every release
2. Readme badges are not working for me on [PyPI](https://pypi.org/project/surfinpy/) and the links are not rendered (PyPI uses [reStructuredText](https://pypi.org/help/#description-content-type))
