### 2019-02-09 5d72408
Fix logo in readthedocs
