+++ docs/source/index.rst
@@ -1,7 +1,7 @@
 surfinpy
 ========
 
-.. include:: ../README.rst
+.. include:: README.rst
 
 .. toctree::
    :maxdepth: 1
