+++ README.md
@@ -5,14 +5,20 @@
 </p>
 
 [![Documentation Status](https://readthedocs.org/projects/surfinpy/badge/?version=latest)](https://surfinpy.readthedocs.io/en/latest/)
+
 [![Build Status](https://travis-ci.com/symmy596/SurfinPy.svg?branch=master)](https://travis-ci.com/symmy596/SurfinPy)
-<a href='https://coveralls.io/github/symmy596/SurfinPy?branch=master'><img src='https://coveralls.io/repos/github/symmy596/SurfinPy/badge.svg?branch=master' alt='Coverage Status' /></a>
+
+<a href='https://coveralls.io/github/symmy596/SurfinPy?branch=master'>
+<img src='https://coveralls.io/repos/github/symmy596/SurfinPy/badge.svg?branch=master' alt='Coverage Status' /></a>
+
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.2549518.svg)](https://doi.org/10.5281/zenodo.2549518)
+
 ![Version](https://img.shields.io/badge/Version-0.8-blue.svg?maxAge=2592000)
 
 [![PyPI version](https://badge.fury.io/py/surfinpy.svg)](https://badge.fury.io/py/surfinpy)
 
-[![Gitter chat](https://gitter.im/SurfinPy/Lobby)](https://gitter.im/gitterHQ/gitter)
+<a href='https://gitter.im/pylj/Lobby'>
+<img src='https://badges.gitter.im/gitterHQ/gitter.png' alt='Gitter chat' /></a>
 
 This is the documentation for the open-source Python project - `surfinpy`.
 A library designed to facilitate the generation of publication ready surface phase diagrams from ab initio calculations.
