### 2019-02-11 #1210 @symmy596
@bocklund Thank you for taking the time to review the project. I think I have addressed your comments

1) I have added the required packages to setup.py under install_requires
2) Reuter2003 was missing the @ and thus was not being added. I have fixed this.
3) Example notebooks have been run and uploaded.
4) HTML logo is now fixed.
5) I have added contact information under the Contributing section of the README. A user now as the option to email me directly, open an issue on the issue tracker, or discuss their question on gitter.

PyPi - I have added - long_description_content_type="text/markdown” to the setup.py. This allows markdown to be rendered on PyPi. An example of a project where this has been done can be found here - https://github.com/bjmorgan/bsym
To the best of my knowledge, a new release is needed to update the Readme on PyPi so this will be fixed when I create the next release. I think it is probably best if I wait for the review from @mkhorton before releasing another version.

If I have missed anything or misunderstood anything then please get back to me.

Question for @labarba
I have added a figure to the paper that illustrates an example output from the code because I think it makes it much clearer what the code is all about. I am not sure I am allowed to edit a paper post submission so if I am not then I am very sorry and I will revert back to the previous version.

Thanks
