### 2019-02-11 #1210 @labarba
It's fine to add a figure if you decided that improves the paper.
