### 2019-02-11 #1210 @mkhorton
I have a minor visual glitch with example 5 script, where the Wulff shape is off-centered/partially cut off and the legend also cut off (using default requirements from the GitHub repo, Python 3.7, macOS).
