### 2019-02-12 #1210 @symmy596
@mkhorton - This sounds like it is a pymatgen issue opposed to a surfinpy issue. surfinpy is used to calculate the surface energies and then these are used to generate the Wulff plot with pymatgen. I suppose the first thing to check is which version of pymatgen are you using? I have been using Version 2018.11.30
