### 2019-02-18 #1210 @symmy596
@bocklund  - Would you like me to comment the specific commits where the changes where made, sorry I realise that I probably should have done this anyway.
Cheers
