### 2019-02-19 #1210 @mkhorton
I'm taking one last look too, but looks good to me as well, will update my checklist shortly (@symmy596 you were correct that the previous issue I mentioned is with pymatgen/matplotlib, not with surfinpy)
