### 2019-02-19 #1210 @mkhorton
I'm happy to recommend publication, looks very useful! Just checked off the remaining items.

The sole outstanding issue is just that the GitHub release (0.7) doesn't match the PyPi version (0.8) which also needs to be updated to correct the Markdown formatting, but as I understand it this is being addressed. I reviewed both the PyPi version + cloned from source.

I would encourage including some of the gallery figures in the GitHub readme, since they're very compelling.
