### 2019-02-19 #1210 @labarba
Regarding the version, @symmy596, after all revisions are done, and you issue a new tagged release, I can update the version number associated to the paper with a `whedon` command here.
