### 2019-02-19 #1210 @symmy596
Hi

@mkhorton I have fixed the docs version number. I will add a couple of figures to the readme, this is a good suggestion.  Thank you @bocklund and @mkhorton for the time that you put in to the review and for your kind comments.

@labarba - I have addressed all of the comments and updated the PyPi release to version 1.0 and I have fixed the PyPi readme - https://pypi.org/project/surfinpy/
I think that should be everything complete now.

Thank you for your help during this review.
