### 2019-02-19 #1210 @bocklund
@labarba I recommend surfinpy for publication in JOSS
