### 2019-02-19 #1210 @labarba
Minor editorial fixes on paper:

- Thus understanding the state of surface is crucial >> Thus, understanding the state of a surface is crucial [missing comma and article]
- using a well established approach >> well-established (hyphen)
- Once the energy is known under different constants the phase >> comma after "constants"
- and thus a phase diagram >> commas before and after this remark
- considering surface defects e.g. vacancies >> commas before and after "e.g." ... same next line
- contructed >> constructed
- spcies >> species
- e.g. oxygen >> comma after "e.g." and after "oxygen"
- method of Marmier (Marmier & Parker, 2004) >> method of Marmier & Parker (2004) [*]
- generating a surface phase diagrams from DFT data >> delete article "a" _or_ remove the plural in "diagrams"
- in Molinary _et al_ (Molinari et al., 2012) >> in Molinary et al. (2012) [*]
- Marmier _et al_ >> same above [*]
- temperature vs pressure >> vs. (with dot)
- Figure 1: This figure illustrates an example phase diagram >> Figure 1: An example phase diagram (cut the clutter)
- physicist and chemists >> physicists
- jupyter notebooks >> capitalize Jupyter

[*] For the syntax to obtain brackets in the right places, see: https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html
