+++ setup.py
@@ -1,6 +1,6 @@
 __author__ = "Adam R. Symington"
 __copyright__ = "Copyright Adam R.Symington (2019)"
-__version__ = "0.9.1"
+__version__ = "1.0"
 __maintainer__ = "Adam R. Symington"
 __email__ = "ars44s@bath.ac.uk"
 __date__ = "25/01/2019"
@@ -10,15 +10,13 @@ import os
 
 module_dir = os.path.dirname(os.path.abspath(__file__))
 
-with open('README.rst', 'r') as file:
-    long_description = file.read()
-
 if __name__ == "__main__":
     setup(
         name='surfinpy',
-        version='0.9.1',
+        version='1.0',
         description='Surface Phase Diagram Plotting Tools',
-        long_description=long_description,
+        long_description=open(os.path.join(module_dir, 'README.md')).read(),
+        long_description_content_type='text/markdown', 
         url='https://github.com/symmy596/SurfinPy',
         author='Adam R. Symington',
         author_email='ars44@bath.ac.uk',
