+++ docs/source/conf.py
@@ -59,7 +59,7 @@ author = 'Adam R.Symington'
 # built documents.
 #
 # The short X.Y version.
-version = '0.4'
+version = '1.0'
 # The full version, including alpha/beta/rc tags.
 release = version
 
