+++ paper/paper.md
@@ -38,7 +38,7 @@ bibliography: paper.bib
 
 A surface phase diagram is a graphical representation of the different physical states of a surface under different conditions.
 The surface represents the first point of contact between the material and the environment.
-Thus understanding the state of surface is crucial for a wide range of problems in materials science concerning the relationship between
+Thus, understanding the state of surface is crucial for a wide range of problems in materials science concerning the relationship between
 the state of the surface and the surrounding environmental condtions.
 Examples include particle morphologies in solid state batteries [@Canepa2018];
 determining the concentration of adsorbed water at a surface depending on synthesis conditions [@Molinari2012] [@Tegner2017];
@@ -47,24 +47,24 @@ catalytic reactions [@Reuter2003]; or determing the effect of dopants and impuri
 Computational modelling can be used to generate surface phase diagrams from energy minimisation data.
 One common research question is how water adsorption affects the surface and material properties.
 The conventional starting point is to perform a series of energy minimisation calculations with varying concentrations of water on several different slabs.
-From the energies, the surface free energy of each calculation (phase) as a function of temperature and pressure can be calculated using a well established approach [@Molinari2012].
-Once the free energy is known under different constants the phase which is most stable at a specific temperature and pressure and thus a phase diagram can be generated.
+From the energies, the surface free energy of each calculation (phase) as a function of temperature and pressure can be calculated using a well-established approach [@Molinari2012].
+Once the free energy is known under different constants, the phase which is most stable at a specific temperature and pressure, and thus a phase diagram, can be generated.
 
-A further degree of complexity can be introduced by considering surface defects e.g. vacancies or interstitials, or other adsorbants e.g. carbon dioxide.
+A further degree of complexity can be introduced by considering surface defects, e.g., vacancies or interstitials, or other adsorbants, e.g., carbon dioxide.
 Using surface defects as an example, it is important to consider the relationship between the defective surface, the stoichiometric surface and the adsorbing water molecules.
-A surface phase diagram can be contructed as a function of the chemcial potential of the adsorbing spcies (water) and the surface defect
-(e.g. oxygen if oxygen vacancies are being considered). This is done using the method of Marmier [@Marmier2004].
+A surface phase diagram can be constructed as a function of the chemcial potential of the adsorbing species (water) and the surface defect
+(e.g., oxygen, if oxygen vacancies are being considered). This is done using the method of Marmier & Parker (2004)[-@Marmier2004].
 
-![This figure illustrates an example phase diagram as a function of chemical potential (a), and as a function of temperature and pressure (b).\label{fig:example}](Figure_1.png)
+![An example phase diagram as a function of chemical potential (a), and as a function of temperature and pressure (b).\label{fig:example}](Figure_1.png)
 
 # `surfinpy`
 
-`surfinpy` is a Python module for generating a surface phase diagrams from DFT data. 
-It contains two core modules for generating surface phase diagrams using both the methods employed in Molinari *et al* [@Molinari2012] and Marmier *et al* [@Marmier2004].
-These allow fast generation of temperature vs pressure phase diagrams and phase diagrams as a function of chemcial potential of species A and B.
+`surfinpy` is a Python module for generating surface phase diagrams from DFT data. 
+It contains two core modules for generating surface phase diagrams using both the methods employed in Molinari *et al* (2012) [-@Molinari2012] and Marmier *et al* (2004)[-@Marmier2004].
+These allow fast generation of temperature vs. pressure phase diagrams and phase diagrams as a function of chemcial potential of species A and B.
 The plotting classes take the outputs of the calculation modules and generate phase diagrams using `matplotlib`.
[…]
