### 2019-02-19 #1210 @labarba
On line 63:

```It contains two core modules for generating surface phase diagrams using both the methods employed in Molinari *et al* [-@Molinari2012] and Marmier *et al* [-@Marmier2004].```

We don't want a manual `*et al*`, but the proper use of markdown citation that gets the Author (year) format.
