### 2019-02-19 #1210 @labarba
I think you want an "in-text citation" : `@smith04 says blah.`
