+++ paper/latex/paper.tex
@@ -42,6 +42,13 @@ Using surface defects as an example, it is important to consider the relationshi
 A surface phase diagram can be contructed as a function of the chemcial potential of the adsorbing spcies (water) and the surface defect 
 (e.g. oxygen if oxygen vacancies are being considered). This is done using the method of Marmier \cite{Marmier2004}. 
 
+
+\begin{figure}[!htb]
+    \centering\includegraphics[width=6in]{../Figure_1.png}
+    \caption{This figure illustrates an example phase diagram as a function of chemical potential (a), and as a function of temperature and pressure (b).}
+    \label{Figure 1}
+  \end{figure}
+
 \section{surfinpy}
 
 surfinpy is a Python module for generating a surface phase diagrams from DFT data. 
