### 2019-02-20 #1210 @symmy596
@labarba Thank you

Tagged release has been created for version 1.0.1
Archive on Zenodo created and the DOI is - 10.5281/zenodo.2573646

Cheers
