### 2019-02-20 #1210 @labarba
@symmy596 Please edit the Zenodo metadata (title and authors) to match the JOSS paper. Thanks!
