### 2019-02-20 #1210 @labarba
Congratulations, @symmy596, your JOSS paper is now published!

Big thank you to the reviewers, @bocklund, @mkhorton — we couldn't do this without you 🙏
